fix: Update annotation, tags; Move URL to POM; Rearrange POM folder
