+++ tests/lesson-01/03-home.spec.ts
@@ -1,5 +1,5 @@
 import { test, expect } from "@playwright/test";
-import { HomePage } from "./pom/home-page";
+import { HomePage } from "./pom/user/home-page";
 import { JSONReader } from "./utils/json-reader";
 
 test.describe("HOME Module", async () => {
@@ -8,10 +8,10 @@ test.describe("HOME Module", async () => {
 
     test.beforeEach(async ({ page }) => {
         homePage = new HomePage(page);
-        await homePage.navigateToUrl(process.env.BASE_URL as string);
+        await homePage.navigateToUrl(homePage.homePageUrl);
     })
 
-    test("HOME_001", { tag: '@UI' }, async () => {
+    test("HOME_001", { annotation: { type: 'Module ID', description: 'HOME' }, tag: ['@HOME_001', '@HOME', '@UI'] }, async () => {
         await test.step("Step 1: Verify title", async () => {
             expect(await homePage.getElementText(homePage.titleXpath)).toBe(expectResult.title);
         })
