+++ tests/lesson-01/pom/admin/dashboard-page.ts
@@ -1,5 +1,5 @@
 import { Page } from "@playwright/test";
-import { BasePage } from "./base-page";
+import { BasePage } from "../base-page";
 
 export class DashboardPage extends BasePage {
     dashboardTextXpath: string = "//h1[text()='Dashboard']";
