+++ tests/lesson-01/pom/user/home-page.ts
@@ -1,7 +1,8 @@
 import { Page } from "@playwright/test";
-import { BasePage } from "./base-page";
+import { BasePage } from "../base-page";
 
 export class HomePage extends BasePage {
+    homePageUrl: string = `${process.env.BASE_URL}`;
     titleXpath: string = "//title";
     headingXpath: string = "//p[@class='site-title']/a";
     productQuantityXpath: string = "//p[@class='woocommerce-result-count']";
